+++ src/index.js
@@ -16,5 +16,9 @@ module.exports.configs = {
   recommended: require('./configs/recommended'),
   strict: require('./configs/strict'),
   tailwind: require('./configs/tailwind'),
+  expanded: require('./configs/expanded'),
+  logical: require('./configs/logical'),
+  migration: require('./configs/migration'),
 };
+module.exports.eslint = require('./eslint');
 module.exports.presets = require('./presets');